refactor: move error boundary reset into fallback component
