+++ packages/site-components/src/500.tsx
@@ -1,24 +1,7 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Hero } from '@jpmorganchase/mosaic-components';
-import { useErrorBoundary } from 'react-error-boundary';
-import { useRouter } from 'next/router';
 
 export function Page500() {
-  const router = useRouter();
-  const { resetBoundary } = useErrorBoundary();
-
-  useEffect(() => {
-    const handleRouteChange = () => {
-      resetBoundary();
-    };
-
-    router.events.on('routeChangeComplete', handleRouteChange);
-
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange);
-    };
-  }, [router]);
-
   return (
     <Hero
       description="A 500 error occurred."
